Safari-609.1.11.1/LayoutTests/http/tests/canvas/philip/tests/security.drawImage.canvas.html

<!DOCTYPE html>
<title>Canvas test: security.drawImage.canvas</title>
<script src="../tests.js"></script>
<link rel="stylesheet" href="../tests.css">
<link rel="prev" href="security.drawImage.image.html" title="security.drawImage.image">
<link rel="next" href="security.pattern.create.html" title="security.pattern.create">
<body class="show_output">
<p>
 <a href="security.drawImage.image.html" accesskey="p" title="[p] security.drawImage.image">&lt;</a>
 <a href="index.html">[index]</a>
 <a href="security.pattern.create.html" accesskey="n" title="[n] security.pattern.create">&gt;</a>
<h1><a href="index.security.html">security</a>.<a href="index.security.drawImage.html">drawImage</a>.canvas</h1>
<p class="desc">drawImage of unclean canvas makes the canvas origin-unclean</p>
<div class="refs">References:
<ul>
<li><a href="spec.html#testrefs.security.drawImage.canvas">security.drawImage.canvas</a>

</ul>
</div>

<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
_addTest(function(canvas, ctx) {

var canvas2 = document.createElement('canvas');
canvas2.width = 100;
canvas2.height = 50;
var ctx2 = canvas2.getContext('2d');
ctx2.drawImage(document.getElementById('yellow.png'), 0, 0);
ctx.drawImage(canvas2, 0, 0);
try { var _thrown = false; canvas.toDataURL(); } catch (e) { _thrown = true; } finally { _assert(_thrown, "should throw exception: canvas.toDataURL()"); }
try { var _thrown = false; ctx.getImageData(0, 0, 1, 1); } catch (e) { _thrown = true; } finally { _assert(_thrown, "should throw exception: ctx.getImageData(0, 0, 1, 1)"); }


});
</script>
<img src="http://localhost:8000/canvas/philip/images/yellow.png" id="yellow.png" class="resource">
